addding

+++ ttshoppingcart/src/main/java/com/techtree/ttshoppingcart/repos/Repository.java
@@ -7,8 +7,6 @@ import com.techtree.ttshoppingcart.model.Users;
 
 
 @org.springframework.stereotype.Repository
-public interface Repository extends  JpaRepository<Users,Long> {
+public interface Repository extends  JpaRepository<Users,Integer> {
 	
-
-
 }
